+++ src/main/java/com/example/smartmessaging/dto/vo/CustomerVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class CustomerVO extends BaseVO {
-    private BigDecimal id;
+    private Long id;
     private String phone;
     private String name;
     private String email;
