+++ src/main/java/com/example/smartmessaging/dto/vo/TagVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,7 +12,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class TagVO extends BaseVO {
-    private BigDecimal id;
+    private Long id;
     private String name;
-    private BigDecimal userId;
+    private Long userId;
 }
